Add instruction on a new dataset in README.md
In this commit, we add instuction on how to run WeakIdent on a new
dataset Lotka Volterra Equation (2D) using config_13 and point out the
file dataset-Python/simulate_lotka_volterra_2d_ode_eqn.py for data
simulateion.

Notice that this dataset is too large to be pushed onto Github, for this
reason, we provide the code to simulate the dataset.